requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
streamlit>=1.42.0
sqlalchemy>=2.0.30
psycopg2-binary>=2.9.9
httpx>=0.27.0
pydantic>=2.8.0
python-dotenv>=1.0.1
graphviz>=0.20.3

langgraph>=0.2.0
langchain-core>=0.3.0
